Added barrelling for song details
